restructure files to put entry code inside test dir
the entry methods that take fwgpu::Matrix<T> as inputs and run the
kernel on them are only needed for testing purposes. These are also not
meant to be or likely to be the API that will be used be the users of
the kernels.

This commit moves all the entry code to files inside of
the testing directory and encloses them inside the fwgpu::testing
namespace for the purpose of keeping the primary kernel directory clean.