change internal helper device allocator to take C as input
the previous allocation helper for benchmarks was only taking A and B as
input arguments, and was calculating the C matrix allocation size
itself. This does not suit our needs where gemm is performed only on a
submatrix of C. This commit change the allocator to take C matrix as an
input as well and initializes the GPU matrix with the input C so we can
add tests in the future for submatrix gemm.