update cutlass srgemm to use MinOp epilogue
currently, we are repurposing the cutlass LinearCombination epilogue
operator for simply storing the computed semiring product to the result
matrix. This is incorrect, as we actually have to perform a min with the
source value of the output matrix C as a modified semiring epilogue.
This commit updates the cutlass submodule to the latest version that
suports this, as well as makes the necessary changes to the kernel.