update cutlass version to use user supplied CTA accumulator initializers
tests added in the previous commit show that all the output values of
the result matrix become zero in our semiring gemm kernel with min and
sum as the operators in some cases.

This was happening because in the cases where input gemm problem size
dimentions are not such that it fills a whole CTA tile, and some of the
registers are initialized to zero by default. Zero initialization is
fine for the regular FFMA case, where zero is the identity value for
sum operation, but in our Srgemm case, infinity is the identity value
for min. The default zero value in our case leads to it being the min value
every time, leading to the entire output matrix being zeroed out.

This version of cutlass fixes that issue by using a user provided inital
value of the accumulators.