cmake: add automatic update and check-out of submodules
a very common mistake of library consumers was to fetch the latest
commits to srgemm-gpu but not update the submodules with it by
re-running "git submodule update". This commit adds cmake machinery that
automatically checks if the submodules are checked out, fetches them if
they are not, and makes sure they are up-to-date.