matrix: add support for row major format
to do this at compile time with template and in a way that does not
break exisiting code, this is implemented as two layout defining structs
within the Matrix.hpp file, which are then supplied as template
arguments. The m_rows and m_cols are moved inside the layout structs now
instead, and the m_layout (an instance of either RowMajor or
ColumnMajor) now holds these row and column counts instead of being
direct members of the Matrix itself.

The layout struct being an instance member of the matrix itself makes
indexing calculations really easy, since we do not need some static
struct whose method would need the rows and cols as input parameters,
and allows us to do this index calculation without really fancy and
overly verbose TMP.