rework entire semi-ring gemm implementation as an extension of cutlass
Previous release of semi-ring gemm that was based on cutlass 2.0 relied
heavily on modifying the core cutlass repository itself. This reworked
implementation DOES NOT modify the upstream cutlass 2.3 release, rather
implements the semi-ring gemm as an extension of cutlass templates.

This implementation is much much cleaner than the previous one, and
allows for much more extensibility in the future.