refactor srgemm templates to surface add/mul ops to user instances
even after having separated out the implementation of the tropical
semi-ring as an extension of cutlass rather than a fork, srgemm code was
not truly usable as a library. This is because anyone who wanted to use
custom thread level operators would have had to modify the arch/srmma.h
file itself, and make additions to the cutlass namespace to add the
cutlass::arch::srmma thread level operator and name a Op type for it
such as cutlass::arch::OpSumMin. This is not feasible if we intend to
extend this as a user pluggable library.

On the other hand, accepting device lambdas at run time (as oppossed to compile
time) would be dreadful as the cutlass templates rely heavily on the
ability of the compiler to inline and then optimize the generated asm.

This commit re-architects how we user arch::srmma. Instead of having a
named op-types that dictates which template specialization of arch::srmma we
use, we accept two structs that can be default constructed as template
arguments to all levels of mma starting at the device level interface
(user facing). this is then passed all the way down to a single
arch::srmma which default constructs them. The second requirement for
these structs is that they have a binary () __device__ operator that
accepts two arguments of the same type as the SRGEMM element types and
returns a value of the same type. These two default constructible structs
and their binary operators define the add and multiply reel operations.

In essence, we expose two template operators that arch::srmma just instances.
This eliminates the need for users to modify or add any types in the cutlass
namespace itself. Just pass in the operators as types in the
cutlass::gemm::device::Srgemm template. Many reel operations of interest
are already defined in the cutlass/functional.h file, which further
makes use of this library much easier for users. Implementing custom
reel operations should also require no knowledge of CUDA or cutlass now,
and should take no more than 10 lines of simple C++ code.