remove default arg for register init vals; rename to additive identity
The previous version used to just call initial values of registers
'accum_init_val' which is technically not incorrect, but does not expose
the true semantics of semiring GEMM. This commit renames it to
'additive_identity' to make it more explicit.

This commit also removes the default argument for the additive_identity
because different reels will have different identity values. For the
tropical config specialization, we add using aliases for the add/mul
identities and use the aliased value in the instance code.