generalize semi-ring epilogue to accept arbitrary addition op
The previous epilogue strategy for the semi-ring gemm was the develop a
bespoke version of cutlass::epilogue::thread class for each reel. This
commit generalizes the semi-ring epilogues by accepting an additional
template parameter just like the GEMM kernel that is a default
constructible struct containing a binary () operator that can implement
any reel without having to implement separate epilogue classes. Just
like the previous rework of the GEMM templates, this removes the need
for the user to understand or modify the cutlass namespace instead, and
just use a pre-defined reel op from cutlass/functional.h. Of course,
this does not have to be the same reel as the addition op of the
semiring GEMM itself, but I do not forsee a such a usecase.

This commit also removes the previous, now redundant min_op epilogue,
and replaces the default config of the tropical semiring.

resolves hpcgarage/semiring-gemm#16