add identity and annihilator constants to all reel op structs
This is an extension to hpcgarage/semiring-gemm#15

Having the reel operator structs hold their own identity and annihilator
values makes the interface much more intuitive, prevents users from
having to repeat themselves and reduces the possibilites of making an
error. The device level SRGEMM template no longer needs an additive
identity to be in the input args; Default AdditionOp::Identity is used instead.

This will also massively simplify the interface for the upcoming
cuasr::epilogue::thread::SemiringLinearCombination (part of
hpcgarage/semiring-gemm#16), as the values of alpha and beta
are the multiplicative identity and annihilator. The interface for a
generic linear combination epilogue would be super ugly and would
require supplying the same arguments as the GEMM again.