add python file to auto-generate device srgemm tests
This auto-generator uses default semiring configurates and the list of
reel operators we support for now to generate thousands of test cases
for all the semirings. All the tests together are really heavy on
compile times and also take ages to run, so they are split into three
levels, each with increasing amount of coverage. By default, only the
smallest number of tests are enabled to make dev iteration times small.

continues hpcgarage/semiring-gemm#12