modernize CMake infrastructure
this commit gets rid of the old FindCUDA way of doing things such as
cuda_add_executable etc. and uses the modern CMake native CUDA
integration instead.

We also rework how we set compiler flags to be more "modern".

Added also is the test level version that the users can set with
-DCUASR_TEST_EVEL={0|1|2}

closes hpcgarage/semiring-gemm#18