fix buggy is_source_needed() for semiring liear combination epilogue
The original implementation of the semi-ring version of linear
combination epilogue returned false if epilogue beta was equal
to multiplicative annihilator. This is incorrect, since in
reality, the only time we can avoid computing beta * C
is if the result of mult_op(beta, C) is equal to the identity of the
addition operator, as the result of this multiplication is what is added
to the result of mult_op(alpha, AB).

This bug resulted in the epilogue not actually performing the final add
operation with the value of beta * C. This commit fixes it. All tests
green now!

hpcgarage/semiring-gemm#16