unify scalar and Array<T,N> functional operators into one op-struct
This changes how the core semiring add and multiply core operators are
implemented and greatly simplifies them. We no longer have two separate
templated operator structs, one scalar Op<T> and a second
Op<cutlass::Array<T,N>> that uses the scalar operator within it. This
bifurcated implementation was clunky design because the former is used
in the arch::Srmma within the SRGEMM compute kernel, whereas the latter
is needed for the thread level epilogue operator.

This commit unifies both under a single Op<T, N = 1>, which exposes a
scalar operator that can be used in the arch::Srmma, and 3 more
operators that operator on cutlass::Array<T,N>. N is set to 1 by
default, allowing for the scalar operator to still be used in arch::Srmma
and array operators to be used in the epilogue if N > 1.

Overall this change also reduces the code size, makes the design more
intuitive, and reduces burden on users should they chose to implement
their own custom operators.

improves hpcgarage/semiring-gemm#16