remove addition identity as a run time argument for srgemm
the device and kernel level SRGEMM allowed for an input value that would
be used as the identity of addition operator to initialize the
accumulators with. This is no longer needed, as for SRGEMM, we already
have the template for the operator, and we can use AdditionOp::Identity
at compile time instead to set the initial accumulator values.

The runtime argument for the threadblock scoped SRMMA is preserved however,
and passed in from the kernel scoped SRGEMM, in case this flexibility is
needed for some other use case.